+++ knotx-core/src/main/java/io/knotx/fragments/FragmentContentExtractor.java
@@ -29,7 +29,8 @@ public final class FragmentContentExtractor {
   }
 
   /**
-   * Extracts pure markup from the given {@link Fragment}'s content removing a script tag with Knot.x specific data.
+   * Extracts pure markup from the given {@link Fragment}'s content removing a script tag with
+   * Knot.x specific data.
    *
    * @param fragment that content will be unwrapped.
    * @return markup without Knotx data processing instructions.
@@ -38,44 +39,48 @@ public final class FragmentContentExtractor {
     if (fragment == null) {
       return null;
     }
-
-    String content = fragment.content();
-    return fragment.isRaw() ? content
-        : content.substring(content.indexOf('>') + 1, content.lastIndexOf('<'));
-  }
-
-  /**
-   * Extracts pure markup from the given {@link Fragment}'s content removing a script tag with Knot.x specific data.
-   *
-   * @param fragment that content will be unwrapped.
-   * @return markup without Knotx data processing instructions in a form of jsoup {@link Document}.
-   */
-  public static Document unwrapDocument(Fragment fragment) {
-    if (fragment == null) {
-      return null;
-    }
-    return unwrapDocument(fragment.content());
+    
+    return fragment.isRaw() ? fragment.content() : extractSnippetContent(fragment.content());
   }
 
   /**
    * Extracts pure markup from the given markup.
    *
    * @param content markup that contains Knotx processing instructions.
-   * @return markup without Knotx data processing instructions in a form of jsoup {@link Document}.
+   * @return markup without Knotx data processing instructions in a form of jsoup {@link Element}.
    */
-  public static Document unwrapDocument(String content) {
-    Element scriptTag = Jsoup.parseBodyFragment(content).body().child(0);
-    return Jsoup.parse(scriptTag.unwrap().toString(), "UTF-8", Parser.xmlParser());
+  public static Element unwrapContent(String content) {
+    Document result = null;
+    if (content != null) {
+      final String unwrapped = extractSnippetContent(content);
+      result = Jsoup.parse(unwrapped, "UTF-8", Parser.xmlParser());
+    }
+    return result;
+  }
+
+  /**
+   * Extracts pure markup from the given {@link Fragment}'s content removing a script tag with
+   * Knot.x specific data.
+   *
+   * @param fragment that content will be unwrapped.
+   * @return markup without Knotx data processing instructions in a form of jsoup {@link Element}.
+   */
+  public static Element unwrapFragmentContent(Fragment fragment) {
+    return fragment == null ? null : unwrapContent(fragment.content());
   }
 
   /**
    * Abbreviate fragment content.
    *
-   * @param Fragment content
+   * @param content to abbreviate
    * @return short fragment content used for logging purposes
    */
   public static String abbreviate(String content) {
     return StringUtils.abbreviate(content.replaceAll("[\n\r\t]", ""),
         FragmentConstants.DEBUG_MAX_FRAGMENT_CONTENT_LOG_LENGTH);
+  }
+
+  private static String extractSnippetContent(String content) {
+    return content.substring(content.indexOf('>') + 1, content.lastIndexOf('<'));
   }
 }