+++ knotx-core/src/main/java/io/knotx/launcher/KnotxStarterVerticle.java
@@ -111,7 +111,7 @@ public class KnotxStarterVerticle extends AbstractVerticle {
         .rxDeployVerticle(module.getName(), getModuleOptions(config, module.getAlias()))
         .map(new ModuleDescriptor(module)::setDeploymentId)
         .doOnError(error -> LOGGER.error("Can't deploy {}: {}", module.toDescriptorLine(), error))
-        .onErrorResumeNext(Single::error)
+        .onErrorResumeNext((err) -> Single.just(new ModuleDescriptor(module).setDeploymentFailed()))
         .toObservable();
   }
 
@@ -139,94 +139,66 @@ public class KnotxStarterVerticle extends AbstractVerticle {
     return new StringBuilder(System.lineSeparator())
         .append(
             deployedModules.stream()
-                .map(item -> String.format("\t\tDeployed %s [%s] [%s]", item.getAlias(),
+                .map(item -> String.format("\t\t%s %s [%s] [%s]", statusLine(deployedDescriptor), item.getAlias(),
                     item.getName(), item.getDeploymentId()))
                 .collect(Collectors.joining(System.lineSeparator())))
         .append(System.lineSeparator())
         .toString();
   }
 
+  private String statusLine(ModuleDescriptor descriptor) {
+    String state;
+    if (!descriptor.isDeploymentFailed()) {
+      state = "Deployed";
+    } else {
+      state = "Failed deploying";
+    }
+    return String.format("\t\t%s %s", state, descriptor.toString());
+  }
+
   private void printLogo() {
-    System.out.println(
-        "@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
-    System.out.println(
-        "@@                                  ,,,,,,,,,                                 @@");
-    System.out.println(
-        "@@                                *,,,,,,,,,,,*                               @@");
-    System.out.println(
-        "@@                              @@&,,,,,,,,,,,,,*                             @@");
-    System.out.println(
-        "@@                            @@@@@@,,,,,,,,,,,,,,*                           @@");
-    System.out.println(
-        "@@                          @@@@@@@@@,,,,,,,,,,,,,,,*                         @@");
-    System.out.println(
-        "@@                        @@@@@@@@@@@@/,,,,,,,,,,,,,,,*                       @@");
-    System.out.println(
-        "@@                      @@@@@@@@@@@@@@@#,,,,,,,,,,,,,,,,*                     @@");
-    System.out.println(
-        "@@                    &@@@@@@@@@@@@@@@@@&,,,,,,,,,,,,,,*@@#                   @@");
-    System.out.println(
-        "@@                  @@&,@@@@@@@@@@@@@@@@@@,,,,,,,,,,,,(@@@@@#                 @@");
-    System.out.println(
-        "@@                @@@@&,,%@@@@@@@@@@@@@@@@@*,,,,,,,,,%@@@@@@@@#               @@");
-    System.out.println(
-        "@@              @@@@@@&,,,*@@@@@@@@@@@@@@@@@(,,,,,,,@@@@@@@@@@@@#             @@");
-    System.out.println(
-        "@@            @@@@@@@@&,,,,,@@@@@@@@@@@@@@@@@&,,,,,@@@@@@@@@@@@@@@#           @@");
-    System.out.println(
-        "@@          @@@@@@@@@@&,,,,,,%@@@@@@@@@@@@@@@@@,,/@@@@@@@@@@@@@@@@@,*         @@");
-    System.out.println(
-        "@@        @@@@@@@@@@@@&,,,,,,,*@@@@@@@@@@@@@@@@@%@@@@@@@@@@@@@@@@#,,,,*       @@");
-    System.out.println(
-        "@@      ,@@@@@@@@@@@@@&,,,,,,,,,@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*,,,,,,,*     @@");
-    System.out.println(
-        "@@    ,,,@@@@@@@@@@@@@&,,,,,,,,,,#@@@@@@@@@@@@@@@@@@@@@@@@@@@@@,,,,,,,,,,,*   @@");
-    System.out.println(
-        "@@  *,,,,#@@@@@@@@@@@@&,,,,,,,,,,,*@@@@@@@@@@@@@@@@@@@@@@@@@@(,,,,,,,,,,,,,,* @@");
-    System.out.println(
-        "@@ ,,,,,,*@@@@@@@@@@@@&,,,,,,,,,,,,,@@@@@@@@@@@@@@@@@@@@@@@@,,,,,,,,,,,,,,,,,*@@");
-    System.out.println(
-        "@@,,,,,,,,@@@@@@@@@@@@&,,,,,,,,,,,,,,(@@@@@@@@@@@@@@@@@@@@&,,,,,,,,,,,,,,,,,,,@@");
-    System.out.println(
-        "@@/,,,,,,,&@@@@@@@@@@@&,,,,,,,,,,,,,,,*@@@@@@@@@@@@@@@@@@@/,,,,,,,,,,,,,,,,,,,@@");
-    System.out.println(
-        "@@ *,,,,,,(@@@@@@@@@@@&,,,,,,,,,,,,,,#@@@@@@@@@@@@@@@@@@@@@&,,,,,,,,,,,,,,,,, @@");
-    System.out.println(
-        "@@   ,,,,,*@@@@@@@@@@@&,,,,,,,,,,,,,@@@@@@@@@@@@@@@@@@@@@@@@@,,,,,,,,,,,,,,*  @@");
-    System.out.println(
-        "@@     ,,,,@@@@@@@@@@@&,,,,,,,,,,,*@@@@@@@@@@@@@@@@@@@@@@@@@@@(,,,,,,,,,,*    @@");
-    System.out.println(
-        "@@       ,,,,,,,,,,,,,,,,,,,,,,,,%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@,,,,,,,*      @@");
-    System.out.println(
-        "@@         ,,&@@@@@@@%,,,,,,,,,,@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*,,,*        @@");
-    System.out.println(
-        "@@           @@@@@@@@@@*,,,,,,/@@@@@@@@@@@@@@@@@*@@@@@@@@@@@@@@@@@#*          @@");
-    System.out.println(
-        "@@             @@@@@@@@@,,,,,&@@@@@@@@@@@@@@@@#,,,@@@@@@@@@@@@@@@@            @@");
-    System.out.println(
-        "@@               @@@@@@@,,,,@@@@@@@@@@@@@@@@@*,,,,,%@@@@@@@@@@@@              @@");
-    System.out.println(
-        "@@                 @@@@*,,(@@@@@@@@@@@@@@@@@,,,,,,,,/@@@@@@@@@                @@");
-    System.out.println(
-        "@@                   @,,,&@@@@@@@@@@@@@@@@&,,,,,,,,,,,@@@@@@                  @@");
-    System.out.println(
-        "@@                     *@@@@@@@@@@@@@@@@@(,,,,,,,,,,,,,&@@                    @@");
-    System.out.println(
-        "@@                       @@@@@@@@@@@@@@@*,,,,,,,,,,,,,,*                      @@");
-    System.out.println(
-        "@@                         @@@@@@@@@@@@,,,,,,,,,,,,,,*                        @@");
-    System.out.println(
-        "@@                           @@@@@@@@%,,,,,,,,,,,,,*                          @@");
-    System.out.println(
-        "@@                            @@@@@/,,,,,,,,,,,,*                             @@");
-    System.out.println(
-        "@@                               @@,,,,,,,,,,,,*                              @@");
-    System.out.println(
-        "@@                                 *,,,,,,,,,*                                @@");
-    System.out.println(
-        "@@                                    ,,,*/                                   @@");
-    System.out.println(
-        "@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
+    // @formatter:off
+    System.out.println("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
+    System.out.println("@@                                  ,,,,,,,,,                                 @@");
+    System.out.println("@@                                *,,,,,,,,,,,*                               @@");
+    System.out.println("@@                              @@&,,,,,,,,,,,,,*                             @@");
+    System.out.println("@@                            @@@@@@,,,,,,,,,,,,,,*                           @@");
+    System.out.println("@@                          @@@@@@@@@,,,,,,,,,,,,,,,*                         @@");
+    System.out.println("@@                        @@@@@@@@@@@@/,,,,,,,,,,,,,,,*                       @@");
+    System.out.println("@@                      @@@@@@@@@@@@@@@#,,,,,,,,,,,,,,,,*                     @@");
+    System.out.println("@@                    &@@@@@@@@@@@@@@@@@&,,,,,,,,,,,,,,*@@#                   @@");
+    System.out.println("@@                  @@&,@@@@@@@@@@@@@@@@@@,,,,,,,,,,,,(@@@@@#                 @@");
+    System.out.println("@@                @@@@&,,%@@@@@@@@@@@@@@@@@*,,,,,,,,,%@@@@@@@@#               @@");
+    System.out.println("@@              @@@@@@&,,,*@@@@@@@@@@@@@@@@@(,,,,,,,@@@@@@@@@@@@#             @@");
+    System.out.println("@@            @@@@@@@@&,,,,,@@@@@@@@@@@@@@@@@&,,,,,@@@@@@@@@@@@@@@#           @@");
+    System.out.println("@@          @@@@@@@@@@&,,,,,,%@@@@@@@@@@@@@@@@@,,/@@@@@@@@@@@@@@@@@,*         @@");
+    System.out.println("@@        @@@@@@@@@@@@&,,,,,,,*@@@@@@@@@@@@@@@@@%@@@@@@@@@@@@@@@@#,,,,*       @@");
+    System.out.println("@@      ,@@@@@@@@@@@@@&,,,,,,,,,@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*,,,,,,,*     @@");
+    System.out.println("@@    ,,,@@@@@@@@@@@@@&,,,,,,,,,,#@@@@@@@@@@@@@@@@@@@@@@@@@@@@@,,,,,,,,,,,*   @@");
+    System.out.println("@@  *,,,,#@@@@@@@@@@@@&,,,,,,,,,,,*@@@@@@@@@@@@@@@@@@@@@@@@@@(,,,,,,,,,,,,,,* @@");
+    System.out.println("@@ ,,,,,,*@@@@@@@@@@@@&,,,,,,,,,,,,,@@@@@@@@@@@@@@@@@@@@@@@@,,,,,,,,,,,,,,,,,*@@");
+    System.out.println("@@,,,,,,,,@@@@@@@@@@@@&,,,,,,,,,,,,,,(@@@@@@@@@@@@@@@@@@@@&,,,,,,,,,,,,,,,,,,,@@");
+    System.out.println("@@/,,,,,,,&@@@@@@@@@@@&,,,,,,,,,,,,,,,*@@@@@@@@@@@@@@@@@@@/,,,,,,,,,,,,,,,,,,,@@");
+    System.out.println("@@ *,,,,,,(@@@@@@@@@@@&,,,,,,,,,,,,,,#@@@@@@@@@@@@@@@@@@@@@&,,,,,,,,,,,,,,,,, @@");
+    System.out.println("@@   ,,,,,*@@@@@@@@@@@&,,,,,,,,,,,,,@@@@@@@@@@@@@@@@@@@@@@@@@,,,,,,,,,,,,,,*  @@");
+    System.out.println("@@     ,,,,@@@@@@@@@@@&,,,,,,,,,,,*@@@@@@@@@@@@@@@@@@@@@@@@@@@(,,,,,,,,,,*    @@");
+    System.out.println("@@       ,,,,,,,,,,,,,,,,,,,,,,,,%@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@,,,,,,,*      @@");
+    System.out.println("@@         ,,&@@@@@@@%,,,,,,,,,,@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@*,,,*        @@");
+    System.out.println("@@           @@@@@@@@@@*,,,,,,/@@@@@@@@@@@@@@@@@*@@@@@@@@@@@@@@@@@#*          @@");
+    System.out.println("@@             @@@@@@@@@,,,,,&@@@@@@@@@@@@@@@@#,,,@@@@@@@@@@@@@@@@            @@");
+    System.out.println("@@               @@@@@@@,,,,@@@@@@@@@@@@@@@@@*,,,,,%@@@@@@@@@@@@              @@");
+    System.out.println("@@                 @@@@*,,(@@@@@@@@@@@@@@@@@,,,,,,,,/@@@@@@@@@                @@");
+    System.out.println("@@                   @,,,&@@@@@@@@@@@@@@@@&,,,,,,,,,,,@@@@@@                  @@");
+    System.out.println("@@                     *@@@@@@@@@@@@@@@@@(,,,,,,,,,,,,,&@@                    @@");
+    System.out.println("@@                       @@@@@@@@@@@@@@@*,,,,,,,,,,,,,,*                      @@");
+    System.out.println("@@                         @@@@@@@@@@@@,,,,,,,,,,,,,,*                        @@");
+    System.out.println("@@                           @@@@@@@@%,,,,,,,,,,,,,*                          @@");
+    System.out.println("@@                            @@@@@/,,,,,,,,,,,,*                             @@");
+    System.out.println("@@                               @@,,,,,,,,,,,,*                              @@");
+    System.out.println("@@                                 *,,,,,,,,,*                                @@");
+    System.out.println("@@                                    ,,,*/                                   @@");
+    System.out.println("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
     System.out.println();
+    // @formatter:on
   }
 }