+++ knotx-junit/src/main/java/io/knotx/junit/coercers/KnotxCoercers.java
@@ -19,6 +19,7 @@ import static org.mockito.Mockito.when;
 
 import io.knotx.dataobjects.Fragment;
 import io.knotx.fragments.FragmentConstants;
+import io.knotx.fragments.SnippetPatterns;
 import io.knotx.junit.util.FileReader;
 import io.vertx.core.json.JsonObject;
 import java.io.IOException;
@@ -26,12 +27,23 @@ import org.mockito.Mockito;
 
 public class KnotxCoercers {
 
-  public Fragment provideFragment(String fragmentContentFile) throws IOException {
+  public Fragment provideFragment(String fragmentContentFileWithDefinedSnippetTagName)
+      throws IOException {
+    final String[] params = fragmentContentFileWithDefinedSnippetTagName.split("\\|");
+    final String fragmentContentFile = params[0];
+    final String snippetTagName;
+    if (params.length > 1) {
+      snippetTagName = params[1];
+    } else {
+      snippetTagName = FragmentConstants.DEFAULT_SNIPPET_TAG_NAME;
+    }
     final String fragmentContent = FileReader.readText(fragmentContentFile);
+    final SnippetPatterns patterns = new SnippetPatterns(snippetTagName);
 
     Fragment fragmentMock = Mockito.mock(Fragment.class);
     when(fragmentMock.content()).thenReturn(fragmentContent);
-    when(fragmentMock.isRaw()).thenReturn(!FragmentConstants.ANY_SNIPPET_PATTERN.matcher(fragmentContent).matches());
+    when(fragmentMock.isRaw())
+        .thenReturn(!patterns.getAnySnippetPattern().matcher(fragmentContent).matches());
     return fragmentMock;
   }