+++ knotx-knot/knotx-knot-fragment-assembler/src/main/java/io/knotx/knot/assembler/impl/FragmentAssemblerKnotProxyImpl.java
@@ -17,6 +17,7 @@ package io.knotx.knot.assembler.impl;
 
 import io.knotx.dataobjects.ClientResponse;
 import io.knotx.dataobjects.KnotContext;
+import io.knotx.fragments.SnippetPatterns;
 import io.knotx.knot.AbstractKnotProxy;
 import io.knotx.knot.assembler.FragmentAssemblerOptions;
 import io.netty.handler.codec.http.HttpResponseStatus;
@@ -34,11 +35,13 @@ public class FragmentAssemblerKnotProxyImpl extends AbstractKnotProxy {
 
   private static final Logger LOGGER = LoggerFactory
       .getLogger(FragmentAssemblerKnotProxyImpl.class);
+  private final SnippetPatterns patterns;
 
   private FragmentAssemblerOptions options;
 
   public FragmentAssemblerKnotProxyImpl(FragmentAssemblerOptions options) {
     this.options = options;
+    this.patterns = new SnippetPatterns(options.getSnippetTagName());
   }
 
   @Override
@@ -46,7 +49,7 @@ public class FragmentAssemblerKnotProxyImpl extends AbstractKnotProxy {
     if (hasFragments(knotContext)) {
       try {
         String joinedFragments = knotContext.getFragments().stream()
-            .map(options.getUnprocessedStrategy()::get)
+            .map(fragment -> options.getUnprocessedStrategy().get(fragment, patterns))
             .collect(Collectors.joining());
 
         return Single.just(createSuccessResponse(knotContext, joinedFragments));