+++ knotx-knot/knotx-knot-fragment-assembler/src/test/java/io/knotx/knot/assembler/impl/UnprocessedFragmentStrategyTest.java
@@ -22,6 +22,7 @@ import com.googlecode.zohhak.api.Configure;
 import com.googlecode.zohhak.api.TestWith;
 import com.googlecode.zohhak.api.runners.ZohhakRunner;
 import io.knotx.dataobjects.Fragment;
+import io.knotx.fragments.SnippetPatterns;
 import io.knotx.junit.coercers.KnotxCoercers;
 import io.knotx.junit.util.FileReader;
 import org.junit.runner.RunWith;
@@ -31,36 +32,42 @@ import org.junit.runner.RunWith;
 public class UnprocessedFragmentStrategyTest {
 
   @TestWith({
-      "simple_snippet.txt;simple_snippet.txt",
-      "raw_fragment.txt;raw_fragment.txt"
+      "simple_snippet.txt;simple_snippet.txt;script",
+      "customTag_snippet.txt;customTag_snippet.txt;knotx:snippet",
+      "raw_fragment.txt;raw_fragment.txt;script"
   })
-  public void asIs_whenFragment_expectIgnoredContent(Fragment fragment,
-      String expectedContentFileName) throws Exception {
-    final String unwrappedContent = UnprocessedFragmentStrategy.AS_IS.get(fragment);
+  public void asIs_whenConfiguredSnippetTag_expectIgnoredContent(Fragment fragment,
+      String expectedContentFileName, String snippetTagName) throws Exception {
+    final String unwrappedContent = UnprocessedFragmentStrategy.AS_IS
+        .get(fragment, new SnippetPatterns(snippetTagName));
     final String expectedContent = FileReader.readText(expectedContentFileName);
 
     assertThat(unwrappedContent, equalToIgnoringWhiteSpace(expectedContent));
   }
 
   @TestWith({
-      "simple_snippet.txt;simple_snippet-expected_unwrapped_content.txt",
-      "big_snippet.txt;big_snippet-expected_unwrapped_content.txt"
+      "simple_snippet.txt;simple_snippet-expected_unwrapped_content.txt;script",
+      "customTag_snippet.txt|knotx:snippet;simple_snippet-expected_unwrapped_content.txt;knotx:snippet",
+      "big_snippet.txt;big_snippet-expected_unwrapped_content.txt;script"
   })
-  public void unwrap_withFragment_expectDefinedContentWithComments(Fragment fragment,
-      String expectedContentFileName) throws Exception {
-    final String unwrappedContent = UnprocessedFragmentStrategy.UNWRAP.get(fragment);
+  public void unwrap_whenConfiguredSnippetTag_expectDefinedContentWithComments(Fragment fragment,
+      String expectedContentFileName, String snippetTagName) throws Exception {
+    final String unwrappedContent = UnprocessedFragmentStrategy.UNWRAP
+        .get(fragment, new SnippetPatterns(snippetTagName));
     final String expectedContent = FileReader.readText(expectedContentFileName);
 
     assertThat(unwrappedContent, equalToIgnoringWhiteSpace(expectedContent));
   }
 
   @TestWith({
-      "simple_snippet.txt;simple_snippet-expected_ignored_content.txt",
-      "raw_fragment.txt;raw_fragment.txt" //when fragment is a raw fragment, it is not ignored
+      "simple_snippet.txt;simple_snippet-expected_ignored_content.txt;script",
+      "customTag_snippet.txt;simple_snippet-expected_ignored_content.txt;knotx:snippet",
+      "raw_fragment.txt;raw_fragment.txt;script" //when fragment is a raw fragment, it is not ignored
   })
-  public void ignore_whenFragment_expectIgnoredContent(Fragment fragment,
-      String expectedContentFileName) throws Exception {
-    final String unwrappedContent = UnprocessedFragmentStrategy.IGNORE.get(fragment);
+  public void ignore_whenConfiguredSnippetTag_expectIgnoredContent(Fragment fragment,
+      String expectedContentFileName, String snippetTagName) throws Exception {
+    final String unwrappedContent = UnprocessedFragmentStrategy.IGNORE
+        .get(fragment, new SnippetPatterns(snippetTagName));
     final String expectedContent = FileReader.readText(expectedContentFileName);
 
     assertThat(unwrappedContent, equalToIgnoringWhiteSpace(expectedContent));