+++ knotx-server/src/main/java/io/knotx/server/configuration/RoutingEntry.java
@@ -21,6 +21,9 @@ import io.vertx.core.json.JsonObject;
 import java.util.HashMap;
 import java.util.Map;
 
+/**
+ * Describes a routing entry of Knot.x Server
+ */
 @DataObject(generateConverter = true, publicConverter = false)
 public class RoutingEntry {
 
@@ -77,37 +80,80 @@ public class RoutingEntry {
     onTransition = new HashMap<>();
   }
 
+  /**
+   * @return Regular expression of HTTP Request path
+   */
   public String getPath() {
     return path;
   }
 
+  /**
+   * Sets the Regular expression of HTTP Request path
+   *
+   * @param path a string with Regexp
+   * @return a reference to this, so the API can be used fluently
+   */
   public RoutingEntry setPath(String path) {
     this.path = path;
     return this;
   }
 
+  /**
+   * @return true if CSRF is enabled for this entry
+   */
   public boolean isCsrf() {
     return csrf;
   }
 
+  /**
+   * Enables/Disabled CSRF support for a given routing entry
+   *
+   * @param csrfEnabled true if enabled
+   * @return a reference to this, so the API can be used fluently
+   */
   public RoutingEntry setCsrf(boolean csrfEnabled) {
     this.csrf = csrfEnabled;
     return this;
   }
 
+  /**
+   * @return event bus address of the Knot that should process the request for a given path
+   */
   public String getAddress() {
     return address;
   }
 
+  /**
+   * Sets the event bus address of the Knot that should process the request for a given path
+   *
+   * @param address a string
+   * @return a reference to this, so the API can be used fluently
+   */
   public RoutingEntry setAddress(String address) {
     this.address = address;
     return this;
   }
 
+  /**
+   *
+   * @return a transition mapping
+   */
   public Map<String, RoutingEntry> getOnTransition() {
     return onTransition;
   }
 
+  /**
+   * Describes routing to addresses of other Knots based on the transition trigger returned from current Knot.
+   * {@code
+   *  "onTransition": {
+   *    "go-a": {},
+   *    "go-b": {}
+   *   }
+   * }
+   *
+   * @param onTransition a map of String-RoutingEntry
+   * @return a reference to this, so the API can be used fluently
+   */
   public RoutingEntry setOnTransition(
       Map<String, RoutingEntry> onTransition) {
     this.onTransition = onTransition;