+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to Knot.x will be documented in this file.
 
 ## Unreleased
 List of changes that are finished but not yet released in any final version.
+ - [PR-376](https://github.com/Cognifide/knotx/pull/376) - Knot.x configurations refactor - Changed the way how configurations and it's defaults are build.
+ - [PR-384](https://github.com/Cognifide/knotx/pull/384) - Introduce Knot.x server backpressure mechanism
+
+## Version 1.2.1
+ - [PR-385](https://github.com/Cognifide/knotx/pull/385) - Fix for [#107](https://github.com/Cognifide/knotx/pull/107) - Support for snippet tags other than `script`
+
+## Version 1.2.0
  - [PR-345](https://github.com/Cognifide/knotx/pull/345) - Update to Vert.x 3.5 and RxJava 2
  - [PR-320](https://github.com/Cognifide/knotx/pull/320) - Added KnotxServer configuration parameter to configure fileUploads folder
  - [PR-347](https://github.com/Cognifide/knotx/pull/347) - Improved error logging on Server routing, repository connector & Http service adapters. Added custom headers configuration for requests/responses, etc.
@@ -21,6 +28,7 @@ List of changes that are finished but not yet released in any final version.
  - [PR-374](https://github.com/Cognifide/knotx/pull/374) - Enable keepAlive connection in http client options
  - [PR-380](https://github.com/Cognifide/knotx/pull/380) - Upgrade jsoup to 1.11.2
  - [PR-379](https://github.com/Cognifide/knotx/pull/379) - Added access logging capabilities to the Knotx HTTP Server. Establish standard configuration of logback logger.
+ - [PR-383](https://github.com/Cognifide/knotx/pull/383) - Fix for [#382](https://github.com/Cognifide/knotx/pull/382) - Unhandled exception if query parameter consists of reserved characters.
 
 ## Version 1.1.2
  - [PR-318](https://github.com/Cognifide/knotx/pull/318) - Knot.x returns exit code `30` in case of missing config